+++ src/component/Caption/Caption.tsx
@@ -1,13 +1,12 @@
-import React, { forwardRef } from "react";
+import React from "react";
 import { SharedCaptionProps } from "../../util";
 import { CaptionLabel } from "./CaptionLabel";
 import { CaptionWrapper } from "./CaptionWrapper/CaptionWrapper";
-import { createPolymorphicComponent } from "@mantine/core";
 
-const _Caption = forwardRef<HTMLParagraphElement, SharedCaptionProps>(({ ...other }, ref) => {
-    return <CaptionLabel ref={ref} {...other} />;
+const _Caption = (({ ...other }: SharedCaptionProps) => {
+    return <CaptionLabel {...other} />;
 }) as any;
 
 _Caption.Wrapper = CaptionWrapper;
 
-export const Caption = createPolymorphicComponent<"p", SharedCaptionProps>(_Caption);
+export const Caption = _Caption;