+++ src/component/Caption/CaptionWrapper/CaptionWrapper.tsx
@@ -1,5 +1,5 @@
-import React, { forwardRef } from "react";
-import { Collapse, createPolymorphicComponent, Stack } from "@mantine/core";
+import React from "react";
+import { Collapse, Stack } from "@mantine/core";
 import { CaptionWrapperProps } from "./CaptionWrapperProps";
 import { CaptionLabel } from "../CaptionLabel";
 
@@ -8,17 +8,15 @@ import { CaptionLabel } from "../CaptionLabel";
  *
  * Used to display a caption beneath a component with controllable visibility.
  */
-const _CaptionWrapper = forwardRef<HTMLParagraphElement, CaptionWrapperProps>(({ children, caption, opened = true, ...other }, ref) => {
+export function CaptionWrapper({ children, caption, opened = true, ...other }: CaptionWrapperProps) {
     return (
         <Stack spacing={0} {...other}>
             {children}
             <Collapse in={opened}>
-                <CaptionLabel ref={ref}>
+                <CaptionLabel>
                     {caption}
                 </CaptionLabel>
             </Collapse>
         </Stack>
     );
-});
-
-export const CaptionWrapper = createPolymorphicComponent<"p", CaptionWrapperProps>(_CaptionWrapper);
+}