+++ src/util/toArrayBuffer.ts
@@ -1,6 +1,0 @@
-/**
- * Convert a [Buffer](https://nodejs.org/api/buffer.html) to an [ArrayBuffer](https://developer.mozilla.org/en-US/docs/Web/JavaScript/Reference/Global_Objects/ArrayBuffer)
- */
-export function toArrayBuffer(buffer: Buffer) {
-    return buffer.buffer.slice(buffer.byteOffset, buffer.byteOffset + buffer.byteLength);
-}