+++ src/util/type/Class.ts
@@ -1,4 +1,0 @@
-/**
- * Generic class type.
- */
-export type Class<T> = new (...args: any[]) => T;