+++ src/util/type/Optional.ts
@@ -1,4 +1,0 @@
-/**
- * Make a single property of an interface optional.
- */
-export type Optional<T, K extends keyof T> = Pick<Partial<T>, K> & Omit<T, K>;